[fix] update addtraining/addworkout
