--- app/src/main/java/com/codaid/trailogandroid/main/add_training/TrainingResult.kt
@@ -1,6 +1,0 @@
-package com.codaid.trailogandroid.main.add_training
-
-class TrainingResult (
-    val success: String? = null,
-    val error: Int? = null
-)
