--- app/src/main/java/com/codaid/trailogandroid/main/add_workout/WorkoutResult.kt
@@ -1,6 +1,0 @@
-package com.codaid.trailogandroid.main.add_workout
-
-class WorkoutResult (
-    val success: String? = null,
-    val error: Int? = null
-)
