+++ app/src/main/res/layouts/activity/layout/activity_add_training.xml
@@ -8,7 +8,7 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     tools:openDrawer="start"
-    tools:context=".main.add_training.AddTrainingActivity">
+    tools:context=".AddTrainingActivity">
 
     <include
         layout="@layout/app_bar_add_training"
