+++ app/src/main/res/layouts/activity/layout/activity_add_workout.xml
@@ -8,7 +8,7 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     tools:openDrawer="start"
-    tools:context=".main.add_workout.AddWorkoutActivity">
+    tools:context=".AddWorkoutActivity">
 
     <include
         layout="@layout/app_bar_add_workout"
