+++ app/src/main/res/layouts/app-bar/layout/app_bar_add_workout.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".main.add_workout.AddWorkoutActivity">
+    tools:context=".AddWorkoutActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
